// packages/datadog-ci/src/commands/synthetics/interfaces.ts
export interface DatadogCIConfig extends APIHelperConfig {
  configPath: string
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface SyntheticsCIConfig extends DatadogCIConfig {}
// ...
export interface DeployTestsCommandConfig extends SyntheticsCIConfig {
  configPath: string
  excludeFields?: string[]
  files: string[]
  publicIds: string[]
  subdomain: string
}